Redesign postfix docker image
The existing docker image used broke in production and figuring
out why was not going well.  The image was also from and "untrusted"
source on docker hub and was designed to auto-restart - something
which should be left to Kubernetes.

This new image is based on the ubuntu image which is much safer, is
custom built so better understood and dies on a postfix crash allowing
kubernetes to handle the failure.