Use AWS simple email service relay
The outgoing IP address for the site got banned as any spam
blocker would view our emails as spam.  By migrating to SES we
can avoid a lot of the overhead of trying to avoid the spam blockers.

This means that all emails must come from an @shadowmail.co.uk
address so the incoming emails look less authentic.